+++ components/pages/contact/CREATE_JIRA_TICKET.gql
@@ -1,0 +1,17 @@
+mutation CREATE_JIRA_TICKET(
+  $messageCategory: JiraTicketCategory!
+  $emailAddress: String!
+  $requestText: String!
+  $displayName: String
+) {
+  createJiraTicket(
+    messageCategory: $messageCategory
+    emailAddress: $emailAddress
+    requestText: $requestText
+    displayName: $displayName
+  ) {
+    _links {
+      web
+    }
+  }
+}
